requirements.txt
requests
wtforms
flask
PyYAML
